reactify autocomplete info
